проверка на права в API Боксов и Марок

+++ boxes-service/src/main/java/ru/rsatu/boxes/rest/BoxController.java
@@ -6,11 +6,16 @@ import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
 import ru.rsatu.boxes.dao.BoxRepository;
 import ru.rsatu.boxes.dao.CarBrandRepository;
+import ru.rsatu.boxes.helpers.UserRole;
 import ru.rsatu.boxes.persistence.Box;
 import ru.rsatu.boxes.persistence.CarBrand;
 import ru.rsatu.boxes.dto.BoxDTO;
 import ru.rsatu.boxes.helpers.DomainToDTOMapper;
+import ru.rsatu.boxes.rest.exception.AccessViolation;
+import ru.rsatu.boxes.rest.exception.BadRequest;
 import ru.rsatu.boxes.rest.exception.ResourceNotFound;
+
+import java.security.Principal;
 
 @RestController
 @RequestMapping("/boxes")
@@ -26,22 +31,47 @@ public class BoxController {
         this.carBrandRepository = carBrandRepository;
     }
 
+    /**
+     * Получить информацию о боксах
+     * Только админ имеет доступ
+     */
     @RequestMapping(method = RequestMethod.GET)
-    public Iterable<BoxDTO> getBoxes() {
+    public Iterable<BoxDTO> getBoxes(Principal auth) {
+        if (!(new UserRole(auth.getName())).isAdmin()) {
+            throw new AccessViolation();
+        }
         return boxDTOMapper.mapMany(boxRepository.findAll());
     }
 
+    /**
+     * Получить информацию о боксе
+     * Только админ имеет доступ
+     */
     @RequestMapping(value="/{boxId}", method = RequestMethod.GET)
-    public BoxDTO getBox(@PathVariable Long boxId) {
-        //TODO проверить существование бокса
-        return boxDTOMapper.mapOne(boxRepository.findOne(boxId));
+    public BoxDTO getBox(Principal auth, @PathVariable Long boxId) throws ResourceNotFound {
+        if (!(new UserRole(auth.getName())).isAdmin()) {
+            throw new AccessViolation();
+        }
+
+        Box box = boxRepository.findOne(boxId);
+
+        if (box == null) {
+            throw new ResourceNotFound(boxId, "Box Not Found");
+        }
+
+        return boxDTOMapper.mapOne(box);
     }
 
     /**
-     TODO только админ должен иметь доступ
+     * Добавить бокс
+     * Только админ имеет доступ
      */
     @RequestMapping(method = RequestMethod.POST)
-    public BoxDTO postBox(@RequestParam Long carBrandId, @RequestParam Long price) {
+    public BoxDTO postBox(Principal auth, @RequestParam Long carBrandId, @RequestParam Long price) {
+        if (!(new UserRole(auth.getName())).isAdmin()) {
+            throw new AccessViolation();
+        }
+
         CarBrand b = carBrandRepository.findOne(carBrandId);
         if (b == null) {
             throw new ResourceNotFound(carBrandId, "Car Brand Not Found");
@@ -55,14 +85,20 @@ public class BoxController {
     }
 
     /**
-     TODO только админ должен иметь доступ
+     * Удалить бокс
+     * Только админ имеет доступ
+     * TODO нельзя удалять бокс, пока в нем машина
      */
     @RequestMapping(value="/{boxId}", method = RequestMethod.DELETE)
-    public ResponseEntity<Boolean> deleteBox(@PathVariable Long boxId) {
+    public ResponseEntity<Boolean> deleteBox(Principal auth, @PathVariable Long boxId) {
+        if (!(new UserRole(auth.getName())).isAdmin()) {
+            throw new AccessViolation();
+        }
+
         try {
             boxRepository.delete(boxId);
         } catch(EmptyResultDataAccessException e){
-            return new ResponseEntity<>(false, HttpStatus.BAD_REQUEST);
+            throw new BadRequest("Can not delete Box");
         }
         return new ResponseEntity<>(true, HttpStatus.OK);
     }
